Refactor article components: update property names for consistency, remove unused namespaces, and adjust date handling in article creation and editing.
