+++ SRC/Web/Components/Features/Articles/ArticleList/GetArticles.cs
@@ -7,12 +7,6 @@
 // Project Name :  Web
 // =======================================================
 
-using Mapster;
-
-using Web.Data.Abstractions;
-using Web.Data.Entities;
-using Web.Data.Models;
-
 namespace Web.Components.Features.Articles.ArticleList;
 
 /// <summary>
